adds route session where all session data is fetched and sent to the frontend
