+++ app.js
@@ -3,9 +3,11 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 const cors = require("cors");
+const mongoose = require('mongoose');
+
 
 var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
+var sessionRouter = require('./routes/session');
 var loginRouter = require('./routes/login');
 
 var app = express();
@@ -19,7 +21,17 @@ app.use(cors());
 
 
 app.use('/', indexRouter);
-app.use('/users', usersRouter);
+app.use('/session', sessionRouter);
 app.use('/login', loginRouter);
+
+
+// Connection to mongodb
+const dbURI = 'mongodb+srv://dbUser:feeble-backend@feeble-backend.vutul.mongodb.net/feeble-backend?retryWrites=true&w=majority';
+mongoose.connect(dbURI, { useNewUrlParser: true, useUnifiedTopology: true })
+ .then((result) => console.log('uppkopplad mot databasen'))
+ .catch((err) => console.log(err));
+
+
+
 
 module.exports = app;
